live: add ws endpoint to push into pipeline
